Keep visibility for the original method inside InterArrival.java, move the content to a separate class for unit test purpose.
