+++ src/org/jitsi/impl/neomedia/rtp/remotebitrateestimator/InterArrival.java
@@ -33,10 +33,6 @@ class InterArrival
 
     private static final Logger logger = Logger.getLogger(InterArrival.class);
 
-    private static long subtractAsUnsignedInt32(long t1, long t2) {
-        return (t1 - t2) & 0xFFFFFFFFL;
-    }
-
     /**
      * webrtc/modules/include/module_common_types.h
      *
@@ -44,19 +40,9 @@ class InterArrival
      * @param prevTimestamp
      * @return
      */
-    static boolean isNewerTimestamp(long timestamp, long prevTimestamp)
+    private static boolean isNewerTimestamp(long timestamp, long prevTimestamp)
     {
-        // Distinguish between elements that are exactly 0x80000000 apart.
-        // If t1>t2 and |t1-t2| = 0x80000000: IsNewer(t1,t2)=true,
-        // IsNewer(t2,t1)=false
-        // rather than having IsNewer(t1,t2) = IsNewer(t2,t1) = false.
-        if (subtractAsUnsignedInt32(timestamp, prevTimestamp) == 0x80000000L)
-        {
-            return timestamp > prevTimestamp;
-        }
-        return
-            timestamp != prevTimestamp
-                && subtractAsUnsignedInt32(timestamp, prevTimestamp) < 0x80000000L;
+        return TimestampUtils.isNewerTimestamp(timestamp, prevTimestamp);
     }
 
     /**
@@ -66,10 +52,9 @@ class InterArrival
      * @param timestamp2
      * @return
      */
-    static long latestTimestamp(long timestamp1, long timestamp2)
+    private static long latestTimestamp(long timestamp1, long timestamp2)
     {
-        return
-            isNewerTimestamp(timestamp1, timestamp2) ? timestamp1 : timestamp2;
+        return TimestampUtils.latestTimestamp(timestamp1, timestamp2);
     }
 
     private boolean burstGrouping;
@@ -258,7 +243,7 @@ class InterArrival
             // interval (32 bits) must be due to reordering. This code is almost
             // identical to that in isNewerTimestamp() in module_common_types.h.
             long timestampDiff
-                    = subtractAsUnsignedInt32(timestamp, currentTimestampGroup.firstTimestamp);
+                    = TimestampUtils.subtractAsUnsignedInt32(timestamp, currentTimestampGroup.firstTimestamp);
 
             return timestampDiff < 0x80000000L;
         }
